+++ delphin/codecs/edspenman.py
@@ -8,6 +8,7 @@ from pathlib import Path
 
 import penman
 
+from delphin.exceptions import PyDelphinException
 from delphin.lnk import Lnk
 from delphin.sembase import (role_priority, property_priority)
 from delphin.eds import (EDS, Node)
@@ -29,7 +30,10 @@ def load(source):
     """
     if not hasattr(source, 'read'):
         source = Path(source).expanduser()
-    graphs = penman.load(source)
+    try:
+        graphs = penman.load(source)
+    except penman.PenmanError as exc:
+        raise PyDelphinException('could not decode with Penman') from exc
     xs = [from_triples(g.triples) for g in graphs]
     return xs
 
@@ -43,7 +47,10 @@ def loads(s):
     Returns:
         a list of EDS objects
     """
-    graphs = penman.loads(s)
+    try:
+        graphs = penman.loads(s)
+    except penman.PenmanError as exc:
+        raise PyDelphinException('could not decode with Penman') from exc
     xs = [from_triples(g.triples) for g in graphs]
     return xs
 
@@ -96,14 +103,22 @@ def dumps(es, properties=True, lnk=True, indent=False):
     to_graph = penman.Graph
     graphs = [to_graph(to_triples(e, properties=properties, lnk=lnk))
               for e in es]
-    return penman.dumps(graphs, indent=indent)
+    try:
+        return penman.dumps(graphs, indent=indent)
+    except penman.PenmanError as exc:
+        raise PyDelphinException('could not encode with Penman') from exc
 
 
 def decode(s):
     """
     Deserialize a EDS object from a EDS-PENMAN string.
     """
-    return from_triples(penman.decode(s).triples())
+    try:
+        g = penman.decode(s)
+    except penman.PenmanError as exc:
+        raise PyDelphinException('could not decode with Penman') from exc
+
+    return from_triples(g.triples)
 
 
 def encode(e, properties=True, lnk=True, indent=False):
@@ -125,7 +140,10 @@ def encode(e, properties=True, lnk=True, indent=False):
         indent = None
     triples = to_triples(e, properties=properties, lnk=lnk)
     g = penman.Graph(triples)
-    return penman.encode(g, indent=indent)
+    try:
+        return penman.encode(g, indent=indent)
+    except penman.PenmanError as exc:
+        raise PyDelphinException('could not encode with Penman') from exc
 
 
 def to_triples(e, properties=True, lnk=True):